pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dogs-breed-identification"
version = "0.1.0"
description = "Dog's breed identification using deep learning (PyTorch)."
authors = [
  { name = "Mobin Yousefi", email = "" }
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["deep-learning", "computer-vision", "pytorch", "kaggle", "dog-breeds"]

dependencies = [
  "torch>=2.0.0",
  "torchvision>=0.15.0",
  "torchaudio>=2.0.0",
  "numpy>=1.23.0",
  "pandas>=2.0.0",
  "scikit-learn>=1.3.0",
  "matplotlib>=3.7.0",
  "tqdm>=4.65.0",
  "Pillow>=10.0.0",
  "pyyaml>=6.0.0",
  "kaggle>=1.6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.0.0",
  "black>=24.0.0",
  "ruff>=0.6.0",
  "mypy>=1.8.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
select = ["E", "F", "I", "B"]
ignore = []

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q"
testpaths = ["tests"]